+++ examples/ai-core/src/registry/setup-registry.ts
@@ -1,13 +1,15 @@
 import { anthropic } from '@ai-sdk/anthropic';
+import { mistral } from '@ai-sdk/mistral';
 import { createOpenAI } from '@ai-sdk/openai';
-import { experimental_createModelRegistry as createModelRegistry } from 'ai';
+import { experimental_createProviderRegistry as createProviderRegistry } from 'ai';
 import dotenv from 'dotenv';
 
 dotenv.config();
 
-export const registry = createModelRegistry({
+export const registry = createProviderRegistry({
   // register provider with prefix and default setup:
   anthropic,
+  mistral,
 
   // register provider with prefix and custom setup:
   openai: createOpenAI({