+++ packages/core/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,21 @@
 
 ### Patch Changes
 
+- Updated dependencies [4728c37f]
+- Updated dependencies [8c49166e]
+  - ai@3.1.36
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [06123501]
+  - ai@3.1.35
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [d25566ac]
 - Updated dependencies [87a5d27e]
   - ai@3.1.34