+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/google-vertex
 
+## 0.0.9
+
+### Patch Changes
+
+- 4728c37f: feat (core): add text embedding model support to provider registry
+- Updated dependencies [7910ae84]
+  - @ai-sdk/provider-utils@0.0.14
+
 ## 0.0.8
 
 ### Patch Changes