+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/react
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [7910ae84]
+  - @ai-sdk/provider-utils@0.0.14
+  - @ai-sdk/ui-utils@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes